requirements.txt
google-cloud==0.33.1
PyAudio==0.2.11
